tools/requirements.txt
arxiv
tqdm
requests
pandas
beautifulsoup4
openai
python-dotenv
zhipuai
zai-sdk